+++ api/dashboard_proj.php
@@ -1,8 +1,9 @@
 <?php
+
 require '../vendor/autoload.php';
 
-use App\Session\Login;
 use App\Entity\ProjMaster;
+use App\Session\Login;
 
 Login::requireLogin();
 
@@ -10,11 +11,11 @@ $listaCampus = [
     'Apucarana',
     'Campo Mourão',
     'Curitiba I (EMBAP)',
-    'Curitiba III (FAP)',
+    'Curitiba II (FAP)',
     'Loanda',
     'Paranaguá',
     'Paranavaí',
-    'União da Vitória'
+    'União da Vitória',
 ];
 
 function defTipoExten($tipoExten)
@@ -30,16 +31,15 @@ function defTipoExten($tipoExten)
     return $tipo[$tipoExten];
 }
 
-
-$dadosProjetos = []; 
-$ano = $_GET['ano'] ?? null;   
+$dadosProjetos = [];
+$ano = $_GET['ano'] ?? null;
 $anoEncerrados = $ano ?? date('Y');
 // echo $anoEncerrados;
 // exit;
 
-$campus = $_GET['campus'] ?? null;   
+$campus = $_GET['campus'] ?? null;
 
-//Preciso separar os wheres pois possuem condicoes diferentes
+// Preciso separar os wheres pois possuem condicoes diferentes
 $whereAno = null;
 $whereEncerrados = null;
 
@@ -57,12 +57,12 @@ $whereEncerrados = "
 
 $qtdPorCampus = [];
 
-//Percorre cada campus com um where diferente
+// Percorre cada campus com um where diferente
 foreach ($listaCampus as $campusNome) {
     $condicoesAno = [];
     $condicoesEncerrados = [];
 
-    //Se tem algum ano no filtro 
+    // Se tem algum ano no filtro
     if ($whereAno) {
         $condicoesAno[] = $whereAno;
     }
@@ -71,31 +71,30 @@ foreach ($listaCampus as $campusNome) {
     $condicoesEncerrados[] = $whereEncerrados;
     $condicoesEncerrados[] = "campus = '{$campusNome}'";
 
-    // echo '<pre>'; 
+    // echo '<pre>';
     //     print_r($condicoesEncerrados);
     // echo '</pre>';
 
-    // echo '<pre>'; 
+    // echo '<pre>';
     //     print_r($condicoesAno);
     // echo '</pre>';
     // exit;
 
-    //Separa os wheres com 'and'
+    // Separa os wheres com 'and'
     $whereFinalAno = implode(' AND ', $condicoesAno);
     $whereFinalEncerrados = implode(' AND ', $condicoesEncerrados);
-    
-    //Cria um array associativo para cada campus correspodenndo com a sua quantidade de propostas
+
+    // Cria um array associativo para cada campus correspodenndo com a sua quantidade de propostas
     $qtdPorCampus[] = [
         'campus' => $campusNome,
-        'total'  => (int) ProjMaster::getQntdRegistros($whereFinalAno),
-        'encerrados' => (int) ProjMaster::getQntdRegistros($whereFinalEncerrados)
+        'total' => (int) ProjMaster::getQntdRegistros($whereFinalAno),
+        'encerrados' => (int) ProjMaster::getQntdRegistros($whereFinalEncerrados),
     ];
 }
 
 $projetos = ProjMaster::getRegistros($whereAno, null, null);
 $qtdProjetos = ProjMaster::getQntdRegistros($whereAno);
 $encerrados = ProjMaster::getQntdRegistros($whereEncerrados);
-
 
 foreach ($projetos as $p) {
     $dadosProjetos[] = [
@@ -104,16 +103,15 @@ foreach ($projetos as $p) {
         'campus' => $p->campus,
         'estado' => $p->estado,
         'tipo_exten' => defTipoExten($p->tipo_exten),
-        'created_at' => $p->created_at
+        'created_at' => $p->created_at,
     ];
 }
 
 echo json_encode([
-    'ano'           => $ano ?: 'todos',
-    'qtdProjetos'   => $qtdProjetos,
+    'ano' => $ano ?: 'todos',
+    'qtdProjetos' => $qtdProjetos,
     'qtdEncerrados' => $encerrados,
-    'qtdPorCampus'  => $qtdPorCampus,
+    'qtdPorCampus' => $qtdPorCampus,
     'dadosProjetos' => $dadosProjetos,
 ]);
 exit;
-
