+++ avaliacoes/includes/_formParcial.php
@@ -7,7 +7,7 @@
     </a>
   </section>
   <hr>
-  <h4 style="text-align: center">ANEXO IV</h4>
+  <h4 style="text-align: center">ANEXO III</h4>
   <h3 class="mt-3" style="text-align: center">RELATÓRIO PARCIAL</h3>
 
    <input type="hidden" name="id_prof" value="<?php echo $obProjeto->id_prof; ?>">
