+++ projetostb/visualizar.php
@@ -91,13 +91,13 @@ switch($t) {
 
 
 
-$anexoIII = [3, 4, 5];
-$anexoIV = [1, 2];
+$anexoII = [3, 4, 5];
+$anexoIII = [1, 2];
 
-if (in_array($t, $anexoIII)) { 
-  $title = 'ANEXO III';
+if (in_array($t, $anexoII)) { 
+  $title = 'ANEXO II';
 } else {
-  $title = 'ANEXO IV';
+  $title = 'ANEXO III';
 }
 
 /*
@@ -354,7 +354,7 @@ c {
   ;
 
 
-  if (in_array($t, $anexoIII)) { 
+  if (in_array($t, $anexoII)) { 
     $acec = ($obProjeto->acec == 'S'? 
     '( x ) Sim<br>( <span> </span><span> </span> ) Não<br>' : 
     '( <span> </span><span> </span> ) Sim<br>( x ) Não<br>');
@@ -411,7 +411,7 @@ entre os tr de baixo
 
   use \App\Entity\Area_temat;
 
-  if (in_array($t, $anexoIII)) { 
+  if (in_array($t, $anexoII)) { 
     $html .= 
     '<table class="time">
       <thead>
@@ -462,7 +462,7 @@ entre os tr de baixo
   </table>'
   ;
 
-  if (in_array($t, $anexoIII)) {
+  if (in_array($t, $anexoII)) {
     $tide = $obProjeto->tide == 'S'? '( x ) Sim <span> </span> <span> </span>( <span> </span><span> </span> ) Não' : '( <span> </span><span> </span> ) Sim <span> </span> <span> </span>( x ) Não';  
     $html .= 
     '<table class="time">
@@ -480,7 +480,7 @@ entre os tr de baixo
         </tr>
      </table>';
 
-  } elseif (in_array($t, $anexoIV)) {
+  } elseif (in_array($t, $anexoIII)) {
     $html .= 
     '<table class="time">
       <thead>
@@ -655,9 +655,9 @@ entre os tr de baixo
   
   $titulo = '';
 
-  if (in_array($t, $anexoIII)) {
+  if (in_array($t, $anexoII)) {
     $titulo ='.  Problema e justificativa da proposta';
-  } elseif (in_array($t, $anexoIV)) {
+  } elseif (in_array($t, $anexoIII)) {
     $titulo ='.  Justificativa da proposta';
   }
 
@@ -669,9 +669,9 @@ entre os tr de baixo
   ;
 
 
-  if (in_array($t, $anexoIII)) {
+  if (in_array($t, $anexoII)) {
     $titulo = '.  Objetivos – Geral e Específicos';
-  } elseif (in_array($t, $anexoIV)) {
+  } elseif (in_array($t, $anexoIII)) {
     $titulo = '.  Objetivos';
   }
