+++ propostas/cadastrar.php
@@ -16,8 +16,8 @@ use App\Session\Login;
 Login::requireLogin();
 $user = Login::getUsuarioLogado();
 
-$anexoIV = [1, 2];
-$anexoIII = [3, 4, 5];
+$anexoIII = [1, 2];
+$anexoII = [3, 4, 5];
 
 $t = $_GET['t'];
 
@@ -132,7 +132,7 @@ if (isset($_POST['titulo'])) {
     }
     $obProjeto->regras = $regra;
 
-    if (in_array($t, $anexoIII)) {
+    if (in_array($t, $anexoII)) {
         if ($_POST['ch_semanal'] == null) {
             $obProjeto->ch_semanal = 0;
         } else {
@@ -158,7 +158,7 @@ if (isset($_POST['titulo'])) {
 
     $obProjeto->contribuicao = $_POST['contribuicao'];
 
-    if (in_array($t, $anexoIV)) {
+    if (in_array($t, $anexoIII)) {
         if ($_POST['ch_total'] == null) {
             $obProjeto->ch_total = 0;
         } else {
@@ -196,7 +196,7 @@ if (isset($_POST['titulo'])) {
     $obProjeto->justificativa = $_POST['justificativa'];
     $obProjeto->cronograma = $_POST['cronograma'];
 
-    if (in_array($t, $anexoIII)) {
+    if (in_array($t, $anexoII)) {
         $obProjeto->referencia = $_POST['referencia'];
     }
 
@@ -295,8 +295,8 @@ echo '
 campusNome = '.$user['ca_nome'].' 
 </script>';
 
-if (in_array($t, $anexoIII)) {
-    include __DIR__.'/includes/formAnexoIII.php';
+if (in_array($t, $anexoII)) {
+    include __DIR__.'/includes/formAnexoII.php';
 
     echo '<script src="cnpq.js"></script>';
 
@@ -308,8 +308,8 @@ var sa = document.querySelector("#cnpq_sarea");
 pegarGA();
   
 </script>';
-} elseif (in_array($t, $anexoIV)) {
-    include __DIR__.'/includes/formAnexoIV.php';
+} elseif (in_array($t, $anexoIII)) {
+    include __DIR__.'/includes/formAnexoIII.php';
 } else {
     header('location: index.php?status=error');
     exit;
