+++ propostas/cadastrarADM.php
@@ -17,8 +17,8 @@ if ($acesso1 + $acesso2 == 0) {
     exit;
 }
 
-$anexoIV = [1, 2];
-$anexoIII = [3, 4, 5];
+$anexoIII = [1, 2];
+$anexoII = [3, 4, 5];
 
 $t = $_GET['t'];
 
@@ -134,7 +134,7 @@ if (isset($_POST['titulo'])) {
     $obProjeto->edt = 0;
     $obProjeto->para_avaliar = 0;
 
-    if (in_array($t, $anexoIII)) {
+    if (in_array($t, $anexoII)) {
         if ($_POST['ch_semanal'] == null) {
             $obProjeto->ch_semanal = 0;
         } else {
@@ -153,7 +153,7 @@ if (isset($_POST['titulo'])) {
 
     $obProjeto->contribuicao = $_POST['contribuicao'];
 
-    if (in_array($t, $anexoIV)) {
+    if (in_array($t, $anexoIII)) {
         if ($_POST['ch_total'] == null) {
             $obProjeto->ch_total = 0;
         } else {
@@ -189,7 +189,7 @@ if (isset($_POST['titulo'])) {
     $obProjeto->justificativa = $_POST['justificativa'];
     $obProjeto->cronograma = $_POST['cronograma'];
 
-    if (in_array($t, $anexoIII)) {
+    if (in_array($t, $anexoII)) {
         $obProjeto->referencia = $_POST['referencia'];
     }
 
@@ -301,8 +301,8 @@ function justName() {
 
 include '../includes/header.php';
 
-if (in_array($t, $anexoIII)) {
-    include __DIR__.'/includes/formAnexoIIIADM.php';
+if (in_array($t, $anexoII)) {
+    include __DIR__.'/includes/formAnexoIIADM.php';
 
     echo '<script src="cnpq.js"></script>';
 
@@ -314,8 +314,8 @@ var sa = document.querySelector("#cnpq_sarea");
 pegarGA();
   
 </script>';
-} elseif (in_array($t, $anexoIV)) {
-    include __DIR__.'/includes/formAnexoIVADM.php';
+} elseif (in_array($t, $anexoIII)) {
+    include __DIR__.'/includes/formAnexoIIIADM.php';
 } else {
     header('location: index.php?status=error');
     exit;
