+++ propostas/includes/formulario.php
@@ -1,8 +1,8 @@
 <main>
 
 <?php
-$anexoIII  = [3, 4, 5]; // Programa / Projeto / Serviço
-$anexoIV = [1, 2];    // Curso / Evento
+$anexoII  = [3, 4, 5]; // Programa / Projeto / Serviço
+$anexoIII = [1, 2];    // Curso / Evento
 $t = $obProjeto->tipo_exten;
 
 ?>
@@ -41,7 +41,7 @@ $t = $obProjeto->tipo_exten;
       </div>
     </div>
 <?php
-    if (in_array($t, $anexoIII)) {
+    if (in_array($t, $anexoII)) {
 ?>
     
       <div class="form-group">
@@ -91,7 +91,7 @@ $t = $obProjeto->tipo_exten;
 <label><h5><?=++$n?>. Classificação do Projeto ou Programa</h5></label>
 
 <?php
-if (in_array($t, $anexoIII)) {
+if (in_array($t, $anexoII)) {
 ?>
 
     <div class="row">
@@ -169,7 +169,7 @@ if (in_array($t, $anexoIII)) {
       </div>
      
 <?php
-  if (in_array($t, $anexoIII)) {
+  if (in_array($t, $anexoII)) {
 ?>
       <div class="col">
         <div class="form-group">
@@ -183,7 +183,7 @@ if (in_array($t, $anexoIII)) {
 ?>
 
 <?php
-  if (in_array($t, $anexoIV)) {
+  if (in_array($t, $anexoIII)) {
 ?>
 
       <div class="col">
@@ -198,7 +198,7 @@ if (in_array($t, $anexoIII)) {
 
 <?php
   }  
-  if (in_array($t, $anexoIII)) {
+  if (in_array($t, $anexoII)) {
 ?>
     <div class="col">
         <div class="form-group">
