+++ propostas/includes/pdf.php
@@ -1,4 +1,5 @@
 <?php
+
 ini_set('display_errors', 1);
 ini_set('display_startup_errors', 1);
 error_reporting(E_ALL);
@@ -125,7 +126,8 @@ $html = '<!DOCTYPE HTML PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3
     <div class="page-number"></div>
   </div>
 
-  <h3>ANEXO III</h3>
+
+  <h3>ANEXO II</h3>
   <h4>FORMULÁRIO PARA ELABORAÇÃO DE PROPOSTAS DE PROGRAMAS,
     PROJETOS OU PRESTAÇÃO DE SERVIÇO</h4>
   <h5>*O responsável pelo preenchimento e encaminhamento é o coordenador da Proposta de Extensão
@@ -311,17 +313,11 @@ $html = '<!DOCTYPE HTML PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3
 
 </html>';
 
-
-
-    
 use Dompdf\Dompdf;
+
 $dompdf = new Dompdf(['enable_remote' => true]);
 
 $dompdf->loadHtml($html);
 $dompdf->setPaper('A4', 'portrait');
 $dompdf->render();
-$dompdf->stream("AE1.pdf");
-
-
-
-
+$dompdf->stream('AE1.pdf');
