+++ relatorio/cadastrarp.php
@@ -70,20 +70,32 @@ if (isset($_POST['atvd_per'])) {
     echo '<pre>';
     print_r($_POST);
     echo '</pre>';
-    exit;
-
+    $int = 0;
+    echo ++$int.'<br>';
     $relatorio->idproj = $obProjeto->id;
+    echo ++$int.'<br>';
     $relatorio->regra = $regras['id'];
+    echo ++$int.'<br>';
     $relatorio->para_avaliar = $obProjeto->para_avaliar;
+    echo ++$int.'<br>';
     $relatorio->periodo_ini = $_POST['periodo_ini'];
+    echo ++$int.'<br>';
     $relatorio->periodo_fim = $_POST['periodo_fim'];
+    echo ++$int.'<br>';
     $relatorio->atvd_per = $_POST['atvd_per'];
+    echo ++$int.'<br>';
     $relatorio->alteracoes = $_POST['alteracoes'];
+    echo ++$int.'<br>';
     $relatorio->atvd_prox_per = $_POST['atvd_prox_per'];
+    echo ++$int.'<br>';
     $relatorio->user = $user['id'];
+    echo ++$int.'<br>';
     $relatorio->tramitar = $_POST['tramitar'];
+    echo ++$int.'<br>';
     $relatorio->visita_tec_qtd = $_POST['visita_tec_qtd'];
+    echo ++$int.'<br>';
     $idRel = $relatorio->cadastrar();
+    echo ++$int.'<br>';
 
     $anexosJS = json_decode($_POST['anexosJS']);
     foreach ($anexosJS as &$anx) {
