+++ relatorio/includes/formParcial.php
@@ -19,7 +19,7 @@ if (isset($editar)) {
     
   </section>
   <hr>
-  <h4 style="text-align: center">ANEXO IV</h4>
+  <h4 style="text-align: center">ANEXO III</h4>
   <h3 class="mt-3" style="text-align: center">RELATÓRIO PARCIAL</h3>
 
    <form name="formAnexo" id="formAnexo" method="POST" enctype="multipart/form-data">
